fix: Course Map 컴포넌트 수정
